requirements.txt
requests>=2.31.0
openai>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
itsdangerous>=2.1.2
# SAML dependencies (opcional - descomentar quando for integrar)
# python3-saml>=1.15.0
# pysaml2>=7.4.2
